Removes utils package
