+++ controllers/kubeobjects/deployment_test.go
@@ -1,4 +1,4 @@
-package utils
+package kubeobjects
 
 import (
 	"os"
